% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aseinfer.R
\name{aseinfer}
\alias{aseinfer}
\title{allele specific expression inference}
\usage{
aseinfer(hyList, dHy, coList, dCo, pe = 3, mdl = "BetaB")
}
\arguments{
\item{hyList}{a list of dataframes which contain hybrid replicates of experiment, each dataframe contains two collums, each represents the expression of one allel. The rownames should be gene name.}

\item{dHy}{a vector of numbers which the ratio of the two allels' DAN content in Hybrid replicates data. dHy=(allel A genomic amount)/(allel B genomic amount)}

\item{coList}{same as the hyList, but it's for the co-culture replicates.}

\item{dCo}{similar to dHy, it's the DNA content ratio for co-culture replicates.}

\item{pe}{parallel environment, how many threads you want to use. defalt is 3.}

\item{mdl}{which model you want to use, the default is:"BetaB", which is a beta-binomial model. Another choise is "B", which is a binomial model.}
}
\value{
Return a dataframe of numbers, each row is the result for each gene, and the collumns are the log of effect of cis/trans and their confidence intervals,log of the over-disperse parameter and the corresponding confidence intervals.
}
\description{
Based on allel specif expression results, give out allle expression difference due to cis effect or trans effect.
}
\examples{
aseinfer(hyList=list(aspHy01,aspHy02,aspHy03,aspHy04,aspHy05,
         aspHy06,aspHy07,aspHy08,aspHy09,aspHy10),
         dHy = c(1,1.1,1.5,1,1.23,1,1.42,1.27,1.3,1.23),
         coList=list(aspCo17,aspCo18,aspCo19,aspCo20,aspCo21),
         dCo=c(1,1.1,1.5,1,1.2),
         pe=3,mdl="BetaB")

aseinfer(hyList=list(aspHy01,aspHy02,aspHy03,aspHy04,aspHy05,
         aspHy06,aspHy07,aspHy08),
         dHy = c(1,1.1,1.5,1,1.23,1,1.42,1.27),
         coList=list(aspCo17,aspCo18,aspCo19,aspCo20),
         dCo=c(1,1.1,1.5,1),
         pe=3,mdl="B")
}

