% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dog_function.R
\name{dog_function}
\alias{dog_function}
\title{A Dog Function}
\usage{
dog_function()
}
\arguments{
\item{love}{Do you love dogs? Defaults to False.}
}
\description{
This function allows you to express your love of dogs.
}
\examples{
dog_function()
}
\keyword{ASEinfer}

