% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tig_function.R
\name{tig_function}
\alias{tig_function}
\title{A Tiger Function}
\usage{
tig_function()
}
\arguments{
\item{love}{Do you love dogs? Defaults to False.}
}
\description{
This function allows you to express your love of dogs.
}
\examples{
tig_function()
}
\keyword{ASEinfer}

